chore: clear file structure, add comments
