+++ packages/client/src/v2-events/features/workqueues/EventOverview/components/EventHistory/index.ts
@@ -1,0 +1,1 @@
+export * from './EventHistory'
